requirements.txt
altair==4.1.0
argon2-cffi==20.1.0
astor==0.8.1
astroid==2.4.2
astropy==4.1
async-generator==1.10
attrs==20.2.0
backcall==0.2.0
base58==2.0.1
bidict==0.21.2
biosppy==0.6.1
bleach==3.2.1
blinker==1.4
boto3==1.16.10
botocore==1.19.10
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
colorama==0.4.4
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
enum-compat==0.0.3
flake8==3.8.4
future==0.18.2
gatspy==0.3
gitdb==4.0.5
GitPython==3.1.11
h5py==3.1.0
hrv-analysis==1.0.3
idna==2.10
importlib-metadata==2.0.0
ipykernel==5.3.4
ipython==7.19.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==5.6.4
jedi==0.17.2
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.17.0
jsonschema==3.2.0
jupyter-client==6.1.7
jupyter-core==4.6.3
jupyterlab-pygments==0.1.2
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
matplotlib==3.3.2
mccabe==0.6.1
mistune==0.8.4
mne==0.21.1
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.2
nolds==0.5.2
nose==1.3.7
notebook==6.1.4
numpy==1.19.3
packaging==20.4
pandas==1.1.4
pandocfilters==1.4.3
parso==0.7.1
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.0.1
plotly==4.12.0
prometheus-client==0.8.0
prompt-toolkit==3.0.8
protobuf==3.13.0
ptyprocess==0.6.0
py-ecg-detectors==1.0.2
pyarrow==2.0.0
pycodestyle==2.6.0
pycparser==2.20
pydeck==0.5.0
pyflakes==2.2.0
Pygments==2.7.2
pylint==2.6.0
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2020.4
PyWavelets==1.1.1
pywin32==300
pywinpty==0.5.7
pyzmq==19.0.2
requests==2.24.0
retrying==1.3.3
s3transfer==0.3.3
scikit-learn==0.23.2
scipy==1.5.3
seaborn==0.11.0
Send2Trash==1.5.0
shortuuid==1.0.1
six==1.15.0
sklearn==0.0
smmap==3.0.4
streamlit==0.70.0
terminado==0.9.1
testpath==0.4.4
threadpoolctl==2.1.0
timeloop==1.0.2
toml==0.10.2
toolz==0.11.1
tornado==6.1
traitlets==5.0.5
typed-ast==1.4.1
tzlocal==2.1
urllib3==1.25.11
validators==0.18.1
watchdog==0.10.3
wcwidth==0.2.5
webencodings==0.5.1
wfdb==3.1.1
widgetsnbextension==3.5.1
wrapt==1.12.1
zipp==3.4.0
